Implement snapshot & add local statemachine(in-memory db)

+++ src/main/java/alg/raft/message/LogEntry.java
@@ -2,9 +2,8 @@ package alg.raft.message;
 
 import alg.raft.enums.EntryType;
 
-// 결국은 이 클래스가 grpc 통해서 전달될거임
 public record LogEntry(
-    int sequence,
+    long sequence,
     long term,
     EntryType type,
     String log
